requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
telethon==1.32.1
supabase==2.0.2
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
requests==2.31.0